+++ lib/core/themes/app_theme.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
+import 'package:flutter/services.dart';
+
+class AppTheme {
+  ///main
+  static Color backgroundColor = const Color(0xffffffff);
+  static Color formFieldBackgroundColor = const Color(0xffffffff);
+  static Color enabledFormFieldBorderColor = const Color(0xff616060);
+  static Color focusedFormFieldBorderColor = const Color(0xff000000);
+  static Color appbarBackgroundColor = Colors.blue;
+  static Color primaryColor = Colors.blue;
+  static Color secondaryColor = const Color(0xffffffff);
+
+  ///selections
+  static Color successColor = Colors.green;
+  static Color dangerColor = Colors.red;
+  static Color infoColor = Colors.blue;
+
+  ///texts
+  static Color textColor_1 = const Color(0xff000000);
+  static Color textColor_2 = const Color(0xffffffff);
+  static Color textCaptionColor = const Color(0xff616060);
+}